requirements.txt
altgraph==0.17.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-axes==7.1.0
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-redis==5.4.0
django_neomodel==0.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
ffmpeg-python==0.2.0
future==1.0.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0
grpcio-status==1.72.0
gunicorn==21.2.0
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
libretranslatepy==2.1.1
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==2.1.2
mss==10.0.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
neo4j==5.19.0
neomodel==5.3.3
numpy==2.2.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pefile==2023.2.7
phonenumbers==9.0.4
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
proglog==0.1.11
proto-plus==1.26.1
protobuf==6.30.2
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytube==15.0.0
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
schedule==1.2.2
selenium==4.28.1
sentry-sdk==2.27.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
translate==3.6.1
translator==0.0.9
trio==0.28.0
trio-websocket==0.11.1
ttkbootstrap==1.12.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
web-browser==0.0.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yt-dlp==2025.3.31